Refactor: modularize game.go by extracting functionality into separate files
