+++ game.go
@@ -21,11 +21,9 @@ import (
 	"fmt"
 	"log"
 	"math"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"reflect"
-	"sort"
 	"sync"
 	"time"
 	"unsafe"
@@ -36,7 +34,6 @@ import (
 	"github.com/goplus/spx/v2/internal/debug"
 	"github.com/goplus/spx/v2/internal/engine"
 	"github.com/goplus/spx/v2/internal/engine/platform"
-	gtime "github.com/goplus/spx/v2/internal/time"
 	"github.com/goplus/spx/v2/internal/timer"
 	"github.com/goplus/spx/v2/internal/ui"
 
@@ -106,9 +103,11 @@ func SetDebug(flags dbgFlags) {
 }
 
 // -------------------------------------------------------------------------------------
+// Core Types
 
 type Shape any
 
+// Game represents the main game instance with all core systems
 type Game struct {
 	baseObj
 	eventSinks
@@ -165,7 +164,6 @@ type Game struct {
 	isCollisionByPixel      bool
 	isAutoSetCollisionLayer bool
 
-	//
 	audioAttenuation float64
 	audioMaxDistance float64
 
@@ -262,6 +260,9 @@ func (p *Game) initSpriteMgr() {
 		p.spriteMgr = newSpriteManager()
 	}
 }
+
+// -------------------------------------------------------------------------------------
+// Game Entry Points
 
 // Gopt_Game_Main is required by XGo compiler as the entry of a .gmx project.
 func Gopt_Game_Main(game Gamer, sprites ...Sprite) {
@@ -682,47 +683,6 @@ func (p *Game) loadIndex(g reflect.Value, proj *projConfig) (err error) {
 	return
 }
 
-func (p *Game) setupBackdrop() {
-	imgW, imgH := p.getCostumeSize()
-	dstW := float64(p.worldWidth_)
-	dstH := float64(p.worldHeight_)
-	imgRadio := (imgW / imgH)
-	worldRadio := (dstW / dstH)
-	// scale image's height to fit world's height
-	isScaleHeight := imgRadio > worldRadio
-
-	switch p.mapMode {
-	case mapModeRepeat:
-		repeatX := dstW / imgW
-		repeatY := dstH / imgH
-		p.setMaterialParamsVec4("repeat_scale", mathf.Vec4{
-			X: repeatX,
-			Y: repeatY,
-			Z: 0,
-			W: 0,
-		}, false)
-	case mapModeFillCut:
-		if isScaleHeight {
-			dstH = dstW / imgRadio
-		} else {
-			dstW = dstH * imgRadio
-		}
-	case mapModeFillRatio:
-		if isScaleHeight {
-			dstW = dstH * imgRadio
-		} else {
-			dstH = dstW / imgRadio
-		}
-	default:
-	}
-
-	scaleX := dstW / imgW
-	scaleY := dstH / imgH
-	p.scale = 1
-	checkUpdateCostume(&p.baseObj)
-	spriteMgr.SetScale(p.syncSprite.GetId(), mathf.NewVec2(scaleX, scaleY))
-}
-
 func (p *Game) endLoad(g reflect.Value, proj *projConfig) (err error) {
 	if debugLoad {
 		log.Println("==> EndLoad")
@@ -753,7 +713,8 @@ func Gopt_Game_Reload(game Gamer, index any) (err error) {
 	return
 }
 
-// -----------------------------------------------------------------------------
+// -------------------------------------------------------------------------------------
+// Stage Setup and Special Shapes
 
 type specsp = map[string]any
 
@@ -847,7 +808,8 @@ var (
 	tySprite = reflect.TypeOf((*Sprite)(nil)).Elem()
 )
 
-// -----------------------------------------------------------------------------
+// -------------------------------------------------------------------------------------
+// Game Loop and Scheduler
 
 func (p *Game) runLoop(cfg *Config) (err error) {
 	if debugLoad {
@@ -862,213 +824,6 @@ func (p *Game) runLoop(cfg *Config) (err error) {
 	return nil
 }
 
-type clicker interface {
-	threadObj
-	doWhenClick(this threadObj)
-	getProxy() *engine.Sprite
-	Visible() bool
-}
-
-func (p *Game) doWhenLeftButtonUp(ev *eventLeftButtonUp) {
-	point := ev.Pos
-	p.inputs.checkTracking(point)
-}
-
-func (p *Game) doWhenLeftButtonDown(ev *eventLeftButtonDown) {
-	point := ev.Pos
-
-	// Detect target sprite for both swipe and click events
-	tempItems := p.getTempShapes()
-	count := len(tempItems)
-
-	var target clicker = nil
-	var targetSprite *SpriteImpl = nil
-	for i := 0; i < count; i++ {
-		item := tempItems[count-i-1]
-		if o, ok := item.(clicker); ok {
-			syncSprite := o.getProxy()
-			if syncSprite != nil && o.Visible() {
-				isClicked := spriteMgr.CheckCollisionWithPoint(syncSprite.GetId(), point, true)
-				if isClicked {
-					target = o
-					// Try to get the SpriteImpl from the clicker
-					if sprite, ok := o.(*SpriteImpl); ok {
-						targetSprite = sprite
-					}
-					break
-				}
-			}
-		}
-	}
-
-	// Start swipe tracking with detected target sprite (can be nil for stage swipes)
-	p.inputs.startTracking(point, targetSprite)
-
-	// add a global click cooldown
-	if !p.inputs.canTriggerClickEvent(inputGlobalClickTimerId) {
-		return
-	}
-
-	if target != nil {
-		syncSprite := target.getProxy()
-		if p.inputs.canTriggerClickEvent(syncSprite.GetId()) {
-			target.doWhenClick(target)
-		}
-	} else {
-		if p.inputs.canTriggerClickEvent(inputStageClickTimerId) {
-			p.sinkMgr.doWhenClick(p)
-		}
-	}
-}
-
-func (p *Game) doWhenMouseMove(ev *eventMouseMove) {
-	// Update current mouse position
-	p.mousePos = ev.Pos
-	// If swipe tracking is active, record the movement
-	p.inputs.onMouseMove(ev.Pos)
-}
-
-// handleEvent dispatches events to their respective handlers
-func (p *Game) handleEvent(ev event) {
-	switch e := ev.(type) {
-	case *eventLeftButtonUp:
-		p.doWhenLeftButtonUp(e)
-	case *eventLeftButtonDown:
-		p.doWhenLeftButtonDown(e)
-	case *eventMouseMove:
-		p.doWhenMouseMove(e)
-	case *eventKeyDown:
-		p.sinkMgr.doWhenKeyPressed(e.Key)
-	case *eventStart:
-		p.sinkMgr.doWhenAwake(nil)
-		p.sinkMgr.doWhenStart()
-	case *eventTimer:
-		p.sinkMgr.doWhenTimer(e.Time)
-	}
-}
-
-func (p *Game) fireEvent(ev event) {
-	select {
-	case p.events <- ev:
-	default:
-		log.Println("Event buffer is full. Skip event:", ev)
-	}
-}
-
-func (p *Game) eventLoop(me coroutine.Thread) int {
-	for {
-		var ev event
-		engine.WaitForChan(p.events, &ev)
-		p.handleEvent(ev)
-	}
-}
-
-// processPendingAudios plays any pending audio for sprites
-func (p *Game) processPendingAudios(items []Shape, tempAudios []string) []string {
-	for _, item := range items {
-		if sprite, ok := item.(*SpriteImpl); ok {
-			engine.Lock()
-			tempAudios = append(tempAudios, sprite.pendingAudios...)
-			sprite.pendingAudios = sprite.pendingAudios[:0]
-			engine.Unlock()
-
-			for _, audio := range tempAudios {
-				sprite.playAudio(audio, false)
-			}
-			tempAudios = tempAudios[:0]
-		}
-	}
-	return tempAudios
-}
-
-// processAnimationEvents handles completed animation events for sprites
-func (p *Game) processAnimationEvents(items []Shape, tempAnimations []string) []string {
-	for _, item := range items {
-		if sprite, ok := item.(*SpriteImpl); ok {
-			engine.Lock()
-			tempAnimations = append(tempAnimations, sprite.donedAnimations...)
-			sprite.donedAnimations = sprite.donedAnimations[:0]
-			engine.Unlock()
-
-			for _, animName := range tempAnimations {
-				sprite.onAnimationDone(animName)
-			}
-			tempAnimations = tempAnimations[:0]
-		}
-	}
-	return tempAnimations
-}
-
-func (p *Game) logicLoop(me coroutine.Thread) int {
-	tempAudios := []string{}
-	tempAnimations := []string{}
-	for {
-		p.camera.onUpdate(gtime.DeltaTime())
-		tempItems := p.getTempShapes()
-		p.spriteMgr.flushActivate()
-
-		tempAudios = p.processPendingAudios(tempItems, tempAudios)
-		tempAnimations = p.processAnimationEvents(tempItems, tempAnimations)
-
-		if targetTimer := timer.CheckTimerEvent(); targetTimer >= 0 {
-			p.fireEvent(&eventTimer{Time: targetTimer})
-		}
-		engine.WaitNextFrame()
-		p.showDebugPanel()
-	}
-}
-
-func (p *Game) inputEventLoop(me coroutine.Thread) int {
-	lastLbtnPressed := false
-	lastMousePos := mathf.Vec2{} // Track last mouse position
-	keyEvents := make([]engine.KeyEvent, 0)
-
-	for {
-		// Check mouse button state
-		curLbtnPressed := inputMgr.GetMouseState(MOUSE_BUTTON_LEFT)
-		if curLbtnPressed != lastLbtnPressed {
-			if lastLbtnPressed {
-				p.fireEvent(&eventLeftButtonUp{Pos: p.mousePos})
-			} else {
-				p.fireEvent(&eventLeftButtonDown{Pos: p.mousePos})
-			}
-		}
-		lastLbtnPressed = curLbtnPressed
-
-		// Check mouse movement
-		// Note: We need to get the actual current mouse position from the engine
-		// For now, we'll use the stored mousePos which should be updated elsewhere
-		curMousePos := inputMgr.GetGlobalMousePos()
-		mathfMousePos := mathf.Vec2{X: float64(curMousePos.X), Y: float64(curMousePos.Y)}
-
-		// Check if mouse moved significantly
-		dx := mathfMousePos.X - lastMousePos.X
-		dy := mathfMousePos.Y - lastMousePos.Y
-		if math.Abs(dx) > mouseMovementThreshold || math.Abs(dy) > mouseMovementThreshold {
-			p.inputs.onMouseMove(mathfMousePos)
-			lastMousePos = mathfMousePos
-		}
-
-		// Handle keyboard events
-		keyEvents = engine.GetKeyEvents(keyEvents)
-		for _, ev := range keyEvents {
-			if ev.IsPressed {
-				p.fireEvent(&eventKeyDown{Key: Key(ev.Id)})
-			} else {
-				p.fireEvent(&eventKeyUp{Key: Key(ev.Id)})
-			}
-		}
-		keyEvents = keyEvents[:0]
-		engine.WaitNextFrame()
-	}
-}
-
-func (p *Game) initEventLoop() {
-	gco.Create(nil, p.eventLoop)
-	gco.Create(nil, p.inputEventLoop)
-	gco.Create(nil, p.logicLoop)
-}
-
 func init() {
 	gco = coroutine.New(engine.OnPanic)
 	engine.SetCoroutines(gco)
@@ -1080,6 +835,7 @@ var (
 
 type threadObj = coroutine.ThreadObj
 
+// SchedNow performs immediate scheduling without timeout check
 func SchedNow() int {
 	if isSchedInMain {
 		if time.Since(mainSchedTime) >= time.Second*mainExecTimeoutSec {
@@ -1158,718 +914,6 @@ func runMain(call func()) {
 	mainSchedTime = time.Now()
 	call()
 	isSchedInMain = false
-}
-
-// -----------------------------------------------------------------------------
-func (p *Game) getWindowSize() mathf.Vec2 {
-	x, y := p.windowSize_()
-	return mathf.NewVec2(float64(x), float64(y))
-}
-
-func (p *Game) windowSize_() (int, int) {
-	if p.windowWidth_ == 0 {
-		p.doWindowSize()
-	}
-	return p.windowWidth_, p.windowHeight_
-}
-
-func (p *Game) doWindowSize() {
-	if p.windowWidth_ == 0 {
-		c := p.costumes[p.costumeIndex_]
-		p.windowWidth_, p.windowHeight_ = c.getSize()
-	}
-}
-
-func (p *Game) worldSize_() (int, int) {
-	if p.worldWidth_ == 0 {
-		p.doWorldSize()
-	}
-	return p.worldWidth_, p.worldHeight_
-}
-
-func (p *Game) doWorldSize() {
-	if p.worldWidth_ == 0 {
-		c := p.costumes[p.costumeIndex_]
-		p.worldWidth_, p.worldHeight_ = c.getSize()
-	}
-}
-
-func (p *Game) touchingPoint(dst *SpriteImpl, x, y float64) bool {
-	return dst.touchPoint(x, y)
-}
-
-func (p *Game) touchingSpriteBy(dst *SpriteImpl, name string) *SpriteImpl {
-	if dst == nil {
-		return nil
-	}
-
-	for _, item := range p.spriteMgr.items {
-		if sp, ok := item.(*SpriteImpl); ok && sp != dst {
-			if sp.name == name && (sp.isVisible && !sp.isDying) {
-				if sp.touchingSprite(dst) {
-					return sp
-				}
-
-			}
-		}
-	}
-	return nil
-}
-
-func (p *Game) objectPos(obj any) (float64, float64) {
-	switch v := obj.(type) {
-	case SpriteName:
-		if sp := p.spriteMgr.findSprite(v); sp != nil {
-			return sp.getXY()
-		}
-		panic("objectPos: sprite not found - " + v)
-	case specialObj:
-		if v == Mouse {
-			return p.getMousePos()
-		}
-	case Pos:
-		if v == Random {
-			worldW, worldH := p.worldSize_()
-			mx, my := rand.Intn(worldW), rand.Intn(worldH)
-			return float64(mx - (worldW >> 1)), float64((worldH >> 1) - my)
-		}
-	case Sprite:
-		return spriteOf(v).getXY()
-	}
-	panic("objectPos: unexpected input")
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) EraseAll() {
-	penMgr.DestroyAllPens()
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) getItems() []Shape {
-	return p.spriteMgr.all()
-}
-
-func (p *Game) addShape(child Shape) {
-	p.spriteMgr.addShape(child)
-}
-
-func (p *Game) addClonedShape(src, clone Shape) {
-	p.spriteMgr.addClonedShape(src, clone)
-}
-
-func (p *Game) removeShape(child Shape) {
-	p.spriteMgr.removeShape(child)
-}
-
-func (p *Game) activateShape(child Shape) {
-	p.spriteMgr.activateShape(child)
-}
-
-func (p *Game) findSprite(name SpriteName) *SpriteImpl {
-	return p.spriteMgr.findSprite(name)
-}
-
-func (p *Game) gotoFront(spr *SpriteImpl) {
-	p.goBackLayers(spr, math.MinInt32)
-}
-
-func (p *Game) gotoBack(spr *SpriteImpl) {
-	p.goBackLayers(spr, math.MaxInt32)
-}
-
-func (p *Game) goBackLayers(spr *SpriteImpl, n int) {
-	p.spriteMgr.goBackLayers(spr, n)
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) BackdropName() string {
-	return p.getCostumeName()
-}
-
-func (p *Game) BackdropIndex() int {
-	return p.getCostumeIndex()
-}
-
-type BackdropName = string
-
-// SetBackdrop func:
-//
-//	SetBackdrop(backdrop) or
-//	SetBackdrop(index) or
-//	SetBackdrop(spx.Next)
-//	SetBackdrop(spx.Prev)
-func (p *Game) setBackdrop(backdrop any, wait bool) {
-	if p.goSetCostume(backdrop) {
-		p.setupBackdrop()
-		p.doWindowSize()
-		p.doWhenBackdropChanged(p.getCostumeName(), wait)
-	}
-}
-
-func (p *Game) SetBackdrop__0(backdrop BackdropName) {
-	p.setBackdrop(backdrop, false)
-}
-
-func (p *Game) SetBackdrop__1(index float64) {
-	p.setBackdrop(index, false)
-}
-
-func (p *Game) SetBackdrop__2(index int) {
-	p.setBackdrop(index, false)
-}
-
-func (p *Game) SetBackdrop__3(action switchAction) {
-	p.setBackdrop(action, false)
-}
-
-func (p *Game) SetBackdropAndWait__0(backdrop BackdropName) {
-	p.setBackdrop(backdrop, true)
-}
-func (p *Game) SetBackdropAndWait__1(index float64) {
-	p.setBackdrop(index, true)
-}
-func (p *Game) SetBackdropAndWait__2(index int) {
-	p.setBackdrop(index, true)
-}
-func (p *Game) SetBackdropAndWait__3(action switchAction) {
-	p.setBackdrop(action, true)
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) KeyPressed(key Key) bool {
-	return inputMgr.GetKey(int64(key))
-}
-
-func (p *Game) MouseX() float64 {
-	return p.mousePos.X
-}
-
-func (p *Game) MouseY() float64 {
-	return p.mousePos.Y
-}
-
-func (p *Game) MousePressed() bool {
-	return inputMgr.MousePressed()
-}
-
-func (p *Game) getMousePos() (x, y float64) {
-	return p.MouseX(), p.MouseY()
-}
-
-func (p *Game) Username() string {
-	panic("todo")
-}
-
-// -----------------------------------------------------------------------------
-func (p *Game) WaitNextFrame() float64 {
-	return engine.WaitNextFrame()
-}
-
-func (p *Game) Wait(secs float64) {
-	engine.Wait(secs)
-}
-
-func (p *Game) Timer() float64 {
-	return timer.Timer()
-}
-
-func (p *Game) ResetTimer() {
-	timer.ResetTimer()
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) Ask(msg any) {
-	msgStr, ok := msg.(string)
-	if !ok {
-		msgStr = fmt.Sprint(msg)
-	}
-	if msgStr == "" {
-		println("ask: msg should not be empty")
-		return
-	}
-	p.ask(false, msgStr, func(answer string) {})
-}
-
-func (p *Game) Answer() string {
-	return p.answerVal
-}
-
-func (p *Game) ask(isSprite bool, question string, callback func(string)) {
-	if p.askPanel == nil {
-		p.askPanel = ui.NewUiAsk()
-		p.addShape(p.askPanel)
-	}
-	hasAnswer := false
-	p.askPanel.Show(isSprite, question, func(msg string) {
-		p.answerVal = msg
-		callback(msg)
-		hasAnswer = true
-	})
-	for {
-		if hasAnswer {
-			break
-		}
-		engine.WaitNextFrame()
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-type EffectKind int
-
-const (
-	ColorEffect EffectKind = iota
-	FishEyeEffect
-	WhirlEffect
-	PixelateEffect
-	MosaicEffect
-	BrightnessEffect
-	GhostEffect
-
-	enumNumOfEffect // max index of enum
-)
-
-var greffNames = []string{
-	ColorEffect:      "color_amount",
-	FishEyeEffect:    "fisheye_amount",
-	WhirlEffect:      "whirl_amount",
-	MosaicEffect:     "uv_amount",
-	PixelateEffect:   "pixleate_amount",
-	BrightnessEffect: "brightness_amount",
-	GhostEffect:      "alpha_amount",
-}
-
-func (kind EffectKind) String() string {
-	return greffNames[kind]
-}
-
-func (p *Game) SetGraphicEffect(kind EffectKind, val float64) {
-	p.baseObj.setGraphicEffect(kind, val)
-}
-
-func (p *Game) ChangeGraphicEffect(kind EffectKind, delta float64) {
-	p.baseObj.changeGraphicEffect(kind, delta)
-}
-
-func (p *Game) ClearGraphicEffects() {
-	p.baseObj.clearGraphicEffects()
-}
-
-// -----------------------------------------------------------------------------
-
-type sound *soundConfig
-
-type SoundName = string
-
-func (p *Game) loadSound(name SoundName) (media sound, err error) {
-	if media, ok := p.sounds.audios[name]; ok {
-		return media, nil
-	}
-
-	if debugLoad {
-		log.Println("==> LoadSound", name)
-	}
-	prefix := "sounds/" + name
-	media = new(soundConfig)
-	if err = loadJson(media, p.fs, prefix+"/index.json"); err != nil {
-		println("loadSound failed:", err.Error())
-		return
-	}
-	media.Path = prefix + "/" + media.Path
-	p.sounds.audios[name] = media
-	return
-}
-
-func (p *Game) playSound(sprite *engine.Sprite, audioId engine.Object, name SoundName, isLoop bool, attenuation, maxDistance float64) soundId {
-	m, err := p.loadSound(name)
-	if err != nil {
-		return invalidSoundId
-	}
-	return p.sounds.play(audioId, m, isLoop, false, sprite.Id, attenuation, maxDistance)
-}
-func (p *Game) playSoundAndWait(sprite *engine.Sprite, audioId engine.Object, name SoundName, attenuation, maxDistance float64) {
-	m, err := p.loadSound(name)
-	if err != nil {
-		return
-	}
-	p.sounds.play(audioId, m, false, true, sprite.Id, attenuation, maxDistance)
-}
-func (p *Game) pauseSound(audioId engine.Object, name SoundName) {
-	m, err := p.loadSound(name)
-	if err != nil {
-		return
-	}
-	p.sounds.pause(audioId, m)
-}
-func (p *Game) resumeSound(audioId engine.Object, name SoundName) {
-	m, err := p.loadSound(name)
-	if err != nil {
-		return
-	}
-	p.sounds.resume(audioId, m)
-}
-func (p *Game) stopSound(audioId engine.Object, name SoundName) {
-	m, err := p.loadSound(name)
-	if err != nil {
-		return
-	}
-	p.sounds.stop(audioId, m)
-}
-func (p *Game) Volume() float64 {
-	p.checkAudioId()
-	return p.sounds.getVolume(p.audioId)
-}
-func (p *Game) Play__0(name SoundName, loop bool) {
-	p.checkAudioId()
-	p.playSound(p.syncSprite, p.audioId, name, loop, 0, defaultAudioMaxDist)
-}
-
-func (p *Game) Play__1(name SoundName) {
-	p.Play__0(name, false)
-}
-
-func (p *Game) PlayAndWait(name SoundName) {
-	p.checkAudioId()
-	p.playSoundAndWait(p.syncSprite, p.audioId, name, 0, defaultAudioMaxDist)
-}
-
-func (p *Game) PausePlaying(name SoundName) {
-	p.checkAudioId()
-	p.pauseSound(p.audioId, name)
-}
-func (p *Game) ResumePlaying(name SoundName) {
-	p.checkAudioId()
-	p.resumeSound(p.audioId, name)
-}
-func (p *Game) StopPlaying(name SoundName) {
-	p.checkAudioId()
-	p.stopSound(p.audioId, name)
-}
-
-func (p *Game) stopSoundInstance(instanceId soundId) {
-	p.sounds.stopInstance(instanceId)
-}
-
-func (p *Game) SetVolume(volume float64) {
-	p.checkAudioId()
-	p.sounds.setVolume(p.audioId, volume)
-}
-
-func (p *Game) ChangeVolume(delta float64) {
-	p.checkAudioId()
-	p.sounds.changeVolume(p.audioId, delta)
-}
-
-func (p *Game) GetSoundEffect(kind SoundEffectKind) float64 {
-	p.checkAudioId()
-	return p.sounds.getEffect(p.audioId, kind)
-}
-func (p *Game) SetSoundEffect(kind SoundEffectKind, value float64) {
-	p.checkAudioId()
-	p.sounds.setEffect(p.audioId, kind, value)
-}
-func (p *Game) ChangeSoundEffect(kind SoundEffectKind, delta float64) {
-	p.checkAudioId()
-	p.sounds.changeEffect(p.audioId, kind, delta)
-}
-func (p *Game) checkAudioId() {
-	if p.audioId == 0 {
-		p.audioId = p.sounds.allocAudio()
-	}
-}
-
-func (p *Game) ClearSoundEffects() {
-	panic("todo")
-}
-
-func (p *Game) StopAllSounds() {
-	p.sounds.stopAll()
-}
-
-func (p *Game) Loudness() float64 {
-	if p.aurec == nil {
-		p.aurec = audiorecord.Open(gco)
-	}
-	return p.aurec.Loudness() * 100
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) doBroadcast(msg string, data any, wait bool) {
-	if debugInstr {
-		log.Println("Broadcast", msg, wait)
-	}
-	p.sinkMgr.doWhenIReceive(msg, data, wait)
-}
-
-func (p *Game) Broadcast__0(msg string) {
-	p.doBroadcast(msg, nil, false)
-}
-func (p *Game) Broadcast__1(msg string, data any) {
-	p.doBroadcast(msg, data, false)
-}
-
-func (p *Game) BroadcastAndWait__0(msg string) {
-	p.doBroadcast(msg, nil, true)
-}
-
-func (p *Game) BroadcastAndWait__1(msg string, data any) {
-	p.doBroadcast(msg, data, true)
-}
-
-// -----------------------------------------------------------------------------
-
-func (p *Game) setStageMonitor(target string, val string, visible bool) {
-	for _, item := range p.spriteMgr.items {
-		if sp, ok := item.(*Monitor); ok && sp.val == val && sp.target == target {
-			sp.setVisible(visible)
-			return
-		}
-	}
-}
-
-func (p *Game) HideVar(name string) {
-	p.setStageMonitor("", getVarPrefix+name, false)
-}
-
-func (p *Game) ShowVar(name string) {
-	p.setStageMonitor("", getVarPrefix+name, true)
-}
-
-func (p *Game) getAllShapes() []Shape {
-	return p.spriteMgr.all()
-}
-
-func (p *Game) getTempShapes() []Shape {
-	return p.spriteMgr.getTempShapes()
-}
-
-// -----------------------------------------------------------------------------
-// Widget
-
-type ShapeGetter interface {
-	getAllShapes() []Shape
-}
-
-// GetWidget_ returns the widget instance with given name. It panics if not found.
-// Instead of being used directly, it is meant to be called by `Gopt_Game_Gopx_GetWidget` only.
-// We extract `GetWidget_` to keep `Gopt_Game_Gopx_GetWidget` simple, which simplifies work in ispx,
-// see details in https://github.com/goplus/builder/issues/765#issuecomment-2313915805.
-func GetWidget_(sg ShapeGetter, name WidgetName) Widget {
-	items := sg.getAllShapes()
-	for _, item := range items {
-		widget, ok := item.(Widget)
-		if ok && widget.GetName() == name {
-			return widget
-		}
-	}
-	panic("GetWidget: widget not found - " + name)
-}
-
-// GetWidget returns the widget instance (in given type) with given name. It panics if not found.
-func Gopt_Game_Gopx_GetWidget[T any](sg ShapeGetter, name WidgetName) *T {
-	widget := GetWidget_(sg, name)
-	if result, ok := widget.(any).(*T); ok {
-		return result
-	} else {
-		panic("GetWidget: type mismatch")
-	}
-}
-
-func (p *Game) checkCollision(ary any) []Sprite {
-	spriteIdAry := ary.([]engine.Object)
-	sprites := make([]Sprite, 1)
-	sort.Slice(spriteIdAry, func(i, j int) bool { return spriteIdAry[i] < spriteIdAry[j] })
-	for _, item := range spriteIdAry {
-		sprite := engine.GetSprite(item)
-		if sprite != nil {
-			impl := sprite.Target.(*SpriteImpl)
-			if impl != nil {
-				sprites = append(sprites, impl.sprite)
-			} else {
-				println(" collision ", item, " not Sprite")
-			}
-		}
-	}
-	return sprites
-}
-
-func (p *Game) IntersectRect(posX, posY, width, height float64) []Sprite {
-	ary := physicMgr.CheckCollisionRect(mathf.NewVec2(posX, posY), mathf.NewVec2(width, height), -1)
-	return p.checkCollision(ary)
-}
-
-func (p *Game) IntersectCircle(posX, posY, radius float64) []Sprite {
-	ary := physicMgr.CheckCollisionCircle(mathf.NewVec2(posX, posY), radius, -1)
-	return p.checkCollision(ary)
-}
-func (p *Game) Raycast__0(fromX, fromY, toX, toY float64, ignoreSprites []Sprite) (hit bool, sprite Sprite, hitX, hitY float64) {
-	from := mathf.NewVec2(fromX, fromY)
-	to := mathf.NewVec2(toX, toY)
-	ignoreSpritesIds := make([]int64, 0)
-	for _, item := range ignoreSprites {
-		if item == nil {
-			continue
-		}
-		impl := spriteOf(item)
-		if impl != nil {
-			ignoreSpritesIds = append(ignoreSpritesIds, impl.getSpriteId())
-		}
-	}
-	result := raycast(from, to, ignoreSpritesIds, -1)
-	if result == nil {
-		return false, nil, 0, 0
-	}
-	var target Sprite = nil
-	if result.Hited {
-		sprite := engine.GetSprite(result.SpriteId)
-		if sprite != nil {
-			impl := sprite.Target.(*SpriteImpl)
-			if impl != nil {
-				target = impl.sprite
-			}
-		}
-	}
-	return result.Hited, target, result.PosX, result.PosY
-}
-
-func (p *Game) Raycast__1(fromX, fromY, toX, toY float64, ignoreSprite Sprite) (hit bool, sprite Sprite, hitX, hitY float64) {
-	return p.Raycast__0(fromX, fromY, toX, toY, []Sprite{ignoreSprite})
-}
-
-func (p *Game) Raycast__2(fromX, fromY, toX, toY float64) (hit bool, sprite Sprite, hitX, hitY float64) {
-	return p.Raycast__0(fromX, fromY, toX, toY, []Sprite{})
-}
-
-func (p *Game) DebugDrawRect(posX, posY, width, height float64, color Color) {
-	debugMgr.DebugDrawRect(mathf.NewVec2(posX, posY), mathf.NewVec2(width, height), toMathfColor(color))
-}
-
-func (p *Game) DebugDrawCircle(posX, posY, radius float64, color Color) {
-	debugMgr.DebugDrawCircle(mathf.NewVec2(posX, posY), radius, toMathfColor(color))
-}
-
-func (p *Game) DebugDrawLine(fromX, fromY, toX, toY float64, color Color) {
-	debugMgr.DebugDrawLine(mathf.NewVec2(fromX, fromY), mathf.NewVec2(toX, toY), toMathfColor(color))
-}
-
-func (p *Game) DebugDrawLines(points []float64, color Color) {
-	if len(points) < 4 || len(points)%2 != 0 {
-		return
-	}
-
-	for i := 0; i < len(points)-2; i += 2 {
-		from := mathf.NewVec2(points[i], points[i+1])
-		to := mathf.NewVec2(points[i+2], points[i+3])
-		debugMgr.DebugDrawLine(from, to, toMathfColor(color))
-	}
-}
-
-// -----------------------------------------------------------------------------
-func (p *Game) setTileMapLayerIndex(index int64) {
-	tilemapMgr.SetLayerIndex(index)
-}
-
-func (p *Game) setTileInfo__0(texturePath string, isCollision bool) {
-	path := engine.ToAssetPath(texturePath)
-	tilemapMgr.SetTile(path, isCollision)
-}
-func (p *Game) setTileInfo__1(texturePath string, collisionPoints []float64) {
-	path := engine.ToAssetPath(texturePath)
-	tilemapMgr.SetTileWithCollisionInfo(path, f64Tof32(collisionPoints))
-}
-
-func (p *Game) PlaceTiles__0(positions []float64, texturePath string) {
-	path := engine.ToAssetPath(texturePath)
-	tilemapMgr.PlaceTiles(f64Tof32(positions), path)
-}
-
-func (p *Game) PlaceTiles__1(positions []float64, texturePath string, layerIndex int64) {
-	path := engine.ToAssetPath(texturePath)
-	tilemapMgr.PlaceTilesWithLayer(f64Tof32(positions), path, layerIndex)
-}
-
-func (p *Game) PlaceTile(x, y float64, texturePath string) {
-	path := engine.ToAssetPath(texturePath)
-	tilemapMgr.PlaceTile(mathf.NewVec2(x, y), path)
-}
-
-func (p *Game) EraseTile__0(x, y float64) {
-	tilemapMgr.EraseTile(mathf.NewVec2(x, y))
-}
-
-func (p *Game) EraseTile__1(x, y float64, layerIndex int64) {
-	tilemapMgr.EraseTileWithLayer(mathf.NewVec2(x, y), layerIndex)
-}
-
-func (p *Game) GetTile__0(x, y float64) string {
-	return tilemapMgr.GetTile(mathf.NewVec2(x, y))
-}
-func (p *Game) GetTile__1(x, y float64, layerIndex int64) string {
-	return tilemapMgr.GetTileWithLayer(mathf.NewVec2(x, y), layerIndex)
-}
-
-func (p *Game) setTileMapOffset(index int64, x, y float64) {
-	tilemapMgr.SetLayerOffset(index, mathf.NewVec2(x, y))
-
-}
-
-func (p *Game) createDecorators(texturePath string, pos mathf.Vec2, rot float64, scale mathf.Vec2, zindex int64, pivot mathf.Vec2) {
-	sceneMgr.CreateStaticSprite(engine.ToAssetPath(texturePath), pos, rot, scale, zindex, pivot, 0, mathf.NewVec2(0, 0), nil)
-}
-
-func (p *Game) createStaticSprite(texturePath string, pos mathf.Vec2, rot float64, scale mathf.Vec2, zindex int64, pivot mathf.Vec2, colliderType string, colliderPivot mathf.Vec2, colliderParams []float64) {
-	colliderTypeInt := paserColliderShapeType(colliderType, 0)
-	sceneMgr.CreateStaticSprite(engine.ToAssetPath(texturePath), pos, rot, scale, zindex, pivot, colliderTypeInt, colliderPivot, colliderParams)
-}
-
-// Path Finding
-func (p *Game) SetupPathFinder__0() {
-	p.setupPathFinder(true, false)
-}
-
-func (p *Game) SetupPathFinder__1(x_grid_size, y_grid_size, x_cell_size, y_cell_size float64, with_jump, with_debug bool) {
-	navigationMgr.SetupPathFinderWithSize(mathf.NewVec2(x_grid_size, y_grid_size), mathf.NewVec2(x_cell_size, y_cell_size), with_jump, with_debug)
-}
-
-func (p *Game) setupPathFinder(with_jump, with_debug bool) {
-	cellSize := mathf.NewVec2(float64(p.pathCellSizeX), float64(p.pathCellSizeY))
-	gridSize := mathf.NewVec2(float64(p.worldWidth_), float64(p.worldHeight_)).Div(cellSize)
-	navigationMgr.SetupPathFinderWithSize(gridSize, cellSize, with_jump, with_debug)
-}
-
-func (p *Game) setObstacle(sprite Sprite, enabled bool) {
-	impl := spriteOf(sprite)
-	if impl != nil {
-		navigationMgr.SetObstacle(impl.getSpriteId(), enabled)
-	}
-}
-
-func (p *Game) FindPath__0(x_from, y_from, x_to, y_to float64) []float64 {
-	return p.FindPath__2(x_from, y_from, x_to, y_to, false, true)
-}
-
-func (p *Game) FindPath__1(x_from, y_from, x_to, y_to float64, with_debug bool) []float64 {
-	return p.FindPath__2(x_from, y_from, x_to, y_to, with_debug, true)
-}
-
-func (p *Game) FindPath__2(x_from, y_from, x_to, y_to float64, with_debug, with_jump bool) []float64 {
-	p.oncePathFinder.Do(func() {
-		p.setupPathFinder(with_jump, with_debug)
-	})
-
-	arr := navigationMgr.FindPath(mathf.NewVec2(x_from, y_from), mathf.NewVec2(x_to, y_to), with_jump)
-	result := arr.([]float32)
-	return f32Tof64(result)
-}
-
-func (p *Game) SetWindowSize(width int64, height int64) {
-	platformMgr.SetWindowSize(width, height, false)
 }
 
 // -----------------------------------------------------------------------------
